Fix for symmetrical option causing sim error
By specifying `Boolean` in the declaration of the option, it caused an error by defaulting to `false`. If you tried to set it to `true` using `--symmetrical=true`, this would result in an error seen on issue  #750. At this time there is still a bug with the symmetric code working.