Issue938 a (#960)
* In addition to the options on a command, there are options on
strategies. There is a callback function in engine.js which sets the
options as defined in a strategy. The strategy objects were not
updated to have a periodSize option, only a period option. So,
the periodSize attribute was not being set. This caused the
execption in this bug.

This commit adds a second option, named 'periodLength'. The name
'periodLength' was chosen, rather than the previous 'periodSize' for
consistency; the description for the --period option already said
'period length', so it made sense that the additional variable be named
that as well.

Each reference to *.periodSize has been changed to *.periodLength.

This commit continues PR #926.

* In addition to the options on a command, there are options on
strategies. There is a callback function in engine.js which sets the
options as defined in a strategy. The strategy objects were not
updated to have a periodSize option, only a period option. So,
the periodSize attribute was not being set. This caused the
execption in this bug.

This commit adds a second option, named 'periodLength'. The name
'periodLength' was chosen, rather than the previous 'periodSize' for
consistency; the description for the --period option already said
'period length', so it made sense that the additional variable be named
that as well.

Each reference to *.periodSize has been changed to *.periodLength.

This commit continues PR #926 and #940. Third time is the charm :)

* #926, #938, #940 - In the backtesters, renaming instances of --period to --periodLength

* #926, #938, #940 - In the backtesters, renaming instances of --period to --periodLength

* Added period and periodLength option, as they were expected for the darwin.js genetic backtest. #926 #938 #940

* #926, #938, #940 - In the (genetic) backtesters, renaming instances of --period to --periodLength