Fix to correct number of days shown in darwin CSV when running with start and end parameters only (#1110)
If you the simulator with no --days parameter and only --start and/or --end, it seems to default the days parameter to 14, not the difference between start and end. This is not serious for the sim because it does it's calculations based on start and end but messes up the CSV output of the darwin backtester by always putting 14 days in the #Days column. This change should fix that.