Fix training start and simulation start/end (#1119)
* Respect buy_max_amt also in paper mode

buy_max_amt was working only on live, now it works also on paper mode

* Fix node-telegram-bot-api deprecated warning

Get rid of that annoying message. Telegram tested and working fine.

* Fix training start and simulation start/end

Previously the training was starting too early if --days_test was
specified. According to the parameter description, --days_test is a
parameter to specify when the simulation should stop (days after
training). The simulator wants the start and end date in this format:
YYYYMMDDhhmm, which was not the case, thus the simulator wasn't
running properly. Now if the user specifies dates it works properly.

Example: ./zenbot.sh train binance.ETH-BTC --start_training 2018-01-01
--end_training 2018-01-03 --days_test 3

In this example, the training stops on Jan 2nd (midnight).
The simulation test runs after the period used in the training, thus
begins on on Jan 3rd and stops 3 days later, on Jan 5th (midnight).

* Ignore models in models/subdirectories