Add option to load previous trades
Closes #1239

This is an opt-in, either use `--use_prev_trades` or set `c.use_prev_trades = true` in conf.js

If enabled we will load N of our previous trades within the time range of all previous trades (currently up to 1000) and use them for initial sell-stop triggers and loss protection.
This way its no longer necessary to wait for a trade in a new session to get triggers/protection to fire, e.g. after a restart.

The age of previous trades is limited so we don't accidentally end up with a stuck bot which isn't trading at all because the market moved too far since our last trade.

All previous trades are exposed by the API in `my_prev_trades` and included in the dashboard under "My Trades" as "muted" trades at the bottom.