Add generation of laugh files for best generation
added launch parameter --generateLaunch

if testing bittrex.ltc-usdt at the end of each generation backtested
will generate launch files gen.bittrex.ltc-usdt.sh and
gen.bittrex.ltc-usdt.bat these files are configured to launch zenbot
using the best overall settings found for that market during the run.
The files will only be overwritten if the current generation found has
a better fitness than the one in the batch files.

the intention for this feature is two fold. 1st make it easier to use
the results of backtested without the copy pasting.  2nd to allow
backtesting to continuously scan the market testing and evolving with
data that comes in always keeping the launch up to date with the best
set, in another processes, it will launch the file and run it for a set
time at the end of set time it will close and relauch with the new
settings, allowing trading strategy to evolve with backtested findings.
a timeout will need to be added to zenbot trade to support this.