Tweak exchange code after long term testing.
Convert cancel call to promise (missed it first time around)  make sure
response is sent back to engine so it accepts that it is canceled.

Catches now pass the error object correctly to the retry handler

getOrder now checks for status of done or closed.
Closed is when the order has been executed on the exchange.  it can
come back as done when the order is partially filled, but then canceled
on the exchange.