Darwin additions (#1404)
Add parameter noStatSave - if true no statistics will be saved for the
generation. useful when using —generateLaunch as it can run for day and
fill the drive with a lot of files.

added some error checking in the process of implementing —noStatSave  —
specifically if a sim fails due to an incompatible parameter
combination (usually TA lib crashes) a bad result is returned that has
a VERY bad fitness calculation.

Saving of generation data now has a more granular format.  instead of a
directory named backtest_YYYMMDDHHmm  it now uses the format
backtes_<exchange>.<market>.YYYMMDDHHmmss
This pervents the condition of overlap when starting multiple darwins
on the same machine within one minute causing them to use the same
simulation directory resulting in errors.