add --ignoreLaunchFitness
previous fitness is ignore and launch files are always updated to the
current best generation as long as the fitness is > 0.

This helps when using —floatScanWindow as over time the best fitness
yesterday may be in reality the worst fitness today.  so the launch
files need to update with the best one, be it greater or less than the
previous, as long as it is profitable.