Library of functions for check of duplicate orders
There is currently a bug in the engine that allows for a buy to occur when a buy is already active or a sell to occur when  a sell is already active these functions check for this condition and return true or false if this condition exists.  see #1629

This check will become a useless step once the problem is corrected in the engine,  as this check will be redundant.