Binance: use real trades instead of OHLCV when asking for trades
In the `getTrades()` method, Binance was previously fetching OHLCV data, in order to generate virtual trades.
The returned result was composed of one (fake) trade per minute, with the direction and volume of the OHLCV data.
Beside, a large part of code was unused and/or contained obvious errors (use of undefined variables, ...)

This commit rewrites the `getTrades()` method to use the real trades returned by Binance.