Avoid returning trades that have been ever returned in bitstamp getTrades() (#1914)
Currently this function returns everything in wstrades array which
doesn't change as long as its length is not larger than 30. This means
if getTrades() gets called multiple times before the length of wstrades
reaches 30, the first couple of trades will be returned every time
getTrades() gets called. This can happen if a short period is used (e.g.
in trend_ema) and if the market is not very active which doesn't make
sense to me.