Stop zenbot from stalling when getTrades is slow (#2334)
`onTrade()` in `engine.js` was returning out of execution immediately when an incoming trade was older than the last known trade. This happens when an exchange API call response is slow than `poll_trades`. By returning immediately out of `onTrade`, further incrementing the trade cursor is prevented, causing the bot to stay in the same time stamp.

Removing the return statement entirely, because further down the function, it's time is already being checked. This way, zenbot resumes operation when an older-than-latest trade comes in.

I welcome any feedback on this, as it's hard to know as a contributor what the design thinking behind this was originally.